Fix up configs to enable debug builds again
