+++ browser/webpack.development.config.js
@@ -1,26 +1,63 @@
-const path = require("path")
+var path = require("path")
+var webpack = require("webpack")
 
-const webpack = require("webpack")
-
-// Override 'development' settings
-const baseConfig = require("./webpack.development.config.js")
-
-const debugConfig = Object.assign({}, baseConfig, {
+module.exports = {
+    entry: [path.join(__dirname, "src/index.tsx")],
+    target: "electron-renderer",
+    externals: {
+        "vscode-jsonrpc": "require('vscode-jsonrpc')",
+        "vscode-textmate": "require('vscode-textmate')",
+        "vscode-languageserver-types": "require('vscode-languageserver-types')",
+        "keyboard-layout": "require('keyboard-layout')",
+        gifshot: "require('gifshot')",
+        "msgpack-lite": "require('msgpack-lite')",
+        "styled-components": "require('styled-components')",
+    },
+    resolve: {
+        extensions: [".tsx", ".ts", ".js", ".less"],
+    },
+    devtool: "cheap-module-eval-source-map",
+    module: {
+        rules: [
+            {
+                test: /\.less$/,
+                use: [
+                    {
+                        loader: "style-loader", // creates style nodes from JS strings
+                    },
+                    {
+                        loader: "css-loader", // translates CSS into CommonJS
+                    },
+                    {
+                        loader: "less-loader", // compiles Less to CSS
+                    },
+                ],
+            },
+            {
+                test: /\.css$/,
+                use: ["style-loader", "css-loader"],
+                exclude: /node_modules/,
+            },
+            {
+                test: /\.tsx?$/,
+                use: "ts-loader",
+                exclude: /node_modules/,
+            },
+        ],
+    },
     plugins: [
         new webpack.DefinePlugin({
             "process.env.NODE_ENV": '"development"',
         }),
-        new webpack.optimize.CommonsChunkPlugin({
-            async: true,
-            minChunks: 2,
-        }),
     ],
     output: {
         path: path.join(__dirname, "..", "lib", "browser"),
-        publicPath: "lib/browser/",
+        publicPath: "http://localhost:8191/",
         filename: "bundle.js",
         chunkFilename: "[name].bundle.js",
     },
-})
-
-module.exports = debugConfig
+    node: {
+        process: false,
+        __dirname: false,
+    },
+}
